backend/apiservice/pyproject.toml
[tool.poetry]
name = "apiservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
aliyun-python-sdk-core = "2.13.16"
aliyun-python-sdk-core-v3 = "2.13.11"
anyio = "3.6.2"
attrs = "19.3.0"
automat = "0.8.0"
blinker = "1.4"
certifi = "2022.9.24"
cffi = "1.15.1"
chardet = "3.0.4"
charset-normalizer = "2.1.1"
click = "7.0"
colorama = "0.4.3"
configobj = "5.0.6"
constantly = "15.1.0"
cryptography = "38.0.3"
distro = "1.4.0"
dnspython = "2.2.1"
ecdsa = "0.18.0"
email-validator = "1.3.0"
entrypoints = "0.3"
fastapi = "0.87.0"
fastapi-jwt-auth = "0.5.0"
h11 = "0.14.0"
httpcore = "0.16.1"
httplib2 = "0.14.0"
httptools = "0.5.0"
httpx = "0.23.1"
hyperlink = "19.0.0"
idna = "3.4"
importlib-metadata = "1.5.0"
incremental = "16.10.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jmespath = "0.10.0"
jsonpatch = "1.22"
jsonpointer = "2.0"
jsonrpcserver = "5.0.9"
jsonschema = "4.0.0"
keyring = "18.0.1"
launchpadlib = "1.10.13"
lazr-restfulclient = "0.14.2"
lazr-uri = "1.0.3"
markupsafe = "2.1.1"
more-itertools = "4.2.0"
oauthlib = "3.1.0"
orjson = "3.8.1"
OSlash = "0.6.3"
passlib = "1.7.4"
pbr = "5.11.0"
pexpect = "4.6.0"
ptyprocess = "0.7.0"
pyasn1 = "0.4.2"
pyasn1-modules = "0.2.1"
pycparser = "2.21"
pydantic = "1.10.2"
PyHamcrest = "1.9.0"
PyJWT = "1.7.1"
pymacaroons = "0.13.0"
pymongo = "4.3.3"
PyNaCl = "1.5.0"
pyOpenSSL = "22.1.0"
pyrsistent = "0.15.5"
pyserial = "3.4"
python-dotenv = "0.21.0"
python-jose = "3.3.0"
python-multipart = "0.0.5"
pywin32-ctypes = "0.2.0"
PyYAML = "5.3.1"
requests = "2.28.1"
requests-unixsocket = "0.2.0"
rfc3986 = "1.5.0"
rsa = "4.9"
SecretStorage = "2.3.1"
service-identity = "18.1.0"
simplejson = "3.16.0"
six = "1.14.0"
sniffio = "1.3.0"
ssh-import-id = "5.10"
starlette = "0.21.0"
testresources = "2.0.1"
typing_extensions = "4.4.0"
ujson = "5.5.0"
urllib3 = "1.26.12"
uvicorn = "0.19.0"
wadllib = "1.3.3"
watchfiles = "0.18.1"
websockets = "10.4"
oss2 = "^2.16.0"
aliyun-python-sdk-dysmsapi = "2.1.2"
eth-account = "^0.8.0"
prometheus-client = "^0.16.0"
fastapi-utils = "^0.2.1"
qrcode = "^7.4.2"
pillow = "^9.5.0"
matplotlib = "^3.7.1"
web3 = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = true
